ESWE-504_service_account_implementation refactored search
